refactor(acp): merge /settings/acp into /settings/agent
The two-route layout (separate /settings/agent for sub-agents and
/settings/acp for ACP config) added navigation complexity that didn't
earn its keep: an acpGated nav item, a disabledByAcp guard on the agent
item, and a redirect block in the settings loader.

With the agent-type dropdown at the top of /settings/agent the same page
serves both modes: OpenHands shows the sub-agents toggle; ACP shows the
preset / command / model form. When ENABLE_ACP is off, the dropdown is
hidden and the page is just the sub-agents settings.

- Merge acp-settings.tsx into agent-settings.tsx (delete acp-settings.tsx).
- Remove route("acp", ...) from routes.ts.
- Remove the acpGated nav entries for /settings/acp from both SAAS and OSS
  nav lists; remove acpGated from the SettingsNavItem interface entirely.
- Remove disabledByAcp from the /settings/agent nav items — the page now
  handles ACP mode rather than being irrelevant to it.
- Remove the acpGated filtering block from use-settings-nav-items.ts.
- Change the disabledByAcp redirect target from /settings/acp →
  /settings/agent (LLM / Condenser / MCP still redirect there when ACP
  is active, since those pages remain irrelevant while ACP is driving).
- Rename acp-settings.test.tsx → agent-settings.test.tsx; update import.
- Fix residual acpGated references in user-context-menu and nav-items tests.
