+++ frontend/__tests__/components/features/user/user-context-menu.test.tsx
@@ -208,7 +208,7 @@ describe("UserContextMenu", () => {
         item.to !== "/settings/billing" &&
         !item.to.startsWith("/settings/org-defaults") &&
         !personalLlmPaths.has(item.to) &&
-        !item.acpGated,
+        true,
     );
 
     await waitFor(() => {
@@ -245,7 +245,7 @@ describe("UserContextMenu", () => {
       (item) =>
         item.to !== "/settings/org-members" &&
         item.to !== "/settings/org" &&
-        !item.acpGated,
+        true,
     );
 
     await waitFor(() => {
@@ -295,7 +295,7 @@ describe("UserContextMenu", () => {
 
       // Wait for the config to load and OSS nav items to appear
       await waitFor(() => {
-        OSS_NAV_ITEMS.filter((item) => !item.acpGated).forEach((item) => {
+        OSS_NAV_ITEMS.forEach((item) => {
           expect(screen.getByText(item.text)).toBeInTheDocument();
         });
       });
